Split image generation logic into own file
